+++ components/layout/PageOptions/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import StyledPageOptions from "./style";
-import Icon from "@components/atoms/Icon";
+import Icon from "@components/shared/Icon";
 import useDarkMode from "use-dark-mode";
 import useLocale from "hooks/useLocale";
