+++ components/shared/Icon/index.js
