+++ components/shared/Icon/style.js
