+++ components/shared/Paragraph/style.js
