+++ pages/index.js
@@ -2,7 +2,7 @@ import About from "@components/home/About";
 import Contact from "@components/home/Contact";
 import Hero from "@components/home/Hero";
 import Projects from "@components/home/Projects";
-import Layout from "@components/Layout";
+import Layout from "@components/layout";
 import locales from "../locales";
 
 export default function MainPage({ content }) {
